refactor: extract core utilities in preparation for `fontless`
