+++ package.json
@@ -119,6 +119,7 @@
       "vue-demi"
     ],
     "onlyBuiltDependencies": [
+      "@tailwindcss/oxide",
       "better-sqlite3",
       "sharp"
     ],
