+++ pnpm-workspace.yaml
@@ -1,8 +1,4 @@
 packages:
-  - "playgrounds/basic"
-  - "playgrounds/scss"
-  - "playgrounds/tailwindcss@3"
-  - "playgrounds/tailwindcss@4"
-  - "playgrounds/unocss"
+  - "playgrounds/*"
   - "docs"
   - "client"
